api/requirements.txt
fastapi
uvicorn
paramiko
Pillow
python-dotenv
httpx